+++ libraries/wuss/test/tasks/ball.c
@@ -34,7 +34,7 @@ result_t ball_create(wuss_t *wuss, const colour_t *palette, ball_task_t *task)
                                    SIZE2D(200, 160),
                                    "Bouncing Ball",
                                    wuss_WINDOW_NONE,
-                                   wuss_NO_BACKGROUND,
+                                   wuss_BACKDROP_COLOUR(wuss_NO_BACKGROUND),
                                    &delegate,
                                    SIZE2D(200, 160),
                                    SIZE2D(0, 0),