+++ libraries/wuss/window/set-background.c
@@ -2,11 +2,11 @@
 
 #include "../impl.h"
 
-result_t wuss_window_set_background(wuss_window_t *window, wuss_colour_t bg)
+result_t wuss_window_set_background(wuss_window_t *window, wuss_backdrop_t bg)
 {
   box_t content;
 
-  if (bg != wuss_NO_BACKGROUND && (bg < 0 || bg >= window->wuss->npalette))
+  if (wuss__validate_backdrop(window->wuss, &bg) != result_OK)
     return result_WUSS_BAD_COLOUR;
 
   window->bg = bg;